refactor: simplify HomePage component and implement session handling
